+++ src/lib/scripts/scripture-reference-utils.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, expect, beforeEach, afterEach, it } from 'vitest';
-import { generateHTMLTest, isBibleBook } from './scripture-reference-utils';
+import { generateHTMLTest, isBibleBook, getReferenceFromString } from './scripture-reference-utils';
 import config from '../../../test_data/data/config';
 import { catalog } from '../../../test_data/data/catalog';
 
@@ -631,6 +631,34 @@ describe('Scripture Reference Utilities', () => {
                 const result = isBibleBook(item, config);
                 expect(result).toEqual(true);
             });
+        });
+    });
+    describe('getReferenceFromString', () => {
+        it('gets book from string', () => {
+            const [collection, book, fromChapter, toChapter, verseRanges]  = getReferenceFromString('MAT 1:1-25');
+            expect(book).toEqual('MAT');
+        });
+        it('gets chapter from string', () => {
+            const [collection, book, fromChapter, toChapter, verseRanges]  = getReferenceFromString('MAT 1:1-25');
+            expect(toChapter).toEqual(1);
+        });
+        it('gets empty book collection when not present', () => {
+            const [collection, book, fromChapter, toChapter, verseRanges]  = getReferenceFromString('MAT 1:1-25');
+            expect(collection).toEqual('');
+        });
+        it('gets verse range from string', () => {
+            const [collection, book, fromChapter, toChapter, verseRanges]  = getReferenceFromString('MAT 1:1-25');
+            const [fromVerse, toVerse, separator] = verseRanges[0];
+            expect(fromVerse).toEqual(1);
+            expect(toVerse).toEqual(25);
+        });
+        it('gets handles chapter only reference', () => {
+            const [collection, book, fromChapter, toChapter, verseRanges]  = getReferenceFromString('MAT 1');
+            const [fromVerse, toVerse, separator] = verseRanges[0];
+            expect(toChapter).toEqual(1);
+            expect(verseRanges.length).toEqual(1);
+            expect(toVerse).toEqual(-1);
+            expect(fromVerse).toEqual(-1);
         });
     });
 });
